+++ Analyzer/src/AnalyzeTopTagger.cc
@@ -43,7 +43,7 @@ void AnalyzeTopTagger::Loop(NTupleReader& tr, double, int maxevents, bool)
         const auto& passMadHT        = tr.getVar<bool>("passMadHT");
         const auto& passTriggerHadMC = tr.getVar<bool>("passTriggerHadMC");
         const auto& NGoodLeptons     = tr.getVar<int>("NGoodLeptons");
-        const auto& HT_trigger_pt45  = tr.getVar<float>("HT_trigger_pt45");
+        const auto& HT_trigger_pt45  = tr.getVar<double>("HT_trigger_pt45");
         const auto& NGoodBJets_pt45  = tr.getVar<int>("NGoodBJets_pt45");
         const auto& NGoodJets_pt45   = tr.getVar<int>("NGoodJets_pt45");
         const auto& dR_bjets         = tr.getVar<float>("dR_bjets");
